Added optional imageAspectRatio property
Without this the content jumps after image is loaded.

The aspect ratio of an image is often "4:3" or "16:9" or similar quite fixed, so it is not a lot of work to use the prop.

The format of the value can be "4:3" or "4/3" or even "4x3". The value is re-formatted as "<number> / <number>".