Added fail-safe background element to angled koros
When card content is high enough and screen is narrow, the background image can be seen above the card.

This is caused by the fixed aspect ratio of the background element which cannot adjust itself to narrow screens when content is high.

The added element is positioned above the content and covers the bg image.