Added shift to the Koros component
Used when the koros shape should overlap the element it is used with.

Creates an illusion of element being masked.

When koros is shifted over an element, sometimes the element must compensate the space the koros takes.

Added a component that matches the taken space: KorosShiftSpacer

Added also a function to get the space in pixels: getShiftHeight. That can be used to add space to css or theme.