Koros adjustments and added element keys
Use compact version of the koros.

Added "bottom-koros-color" to theme.

When using background image the bottom koros should sometimes match the color of the component below Hero.

"koros-color" cannot be used as the koros is merged with the content box in the mobile version.

In desktop version the koros is pointing downwards from the image box.