Convert background-image to img
If background-image was a theme prop and a css variable, reference to image would be calculated from the location of the css file.

And css files can be located in any path. This results in images not found.

Unless absolute urls were used. But that would be too strict requirement.

Removed background images and added img-elements with containers to do the same.