Prevent postcss from converting "inset" css rule
PostCss converts "inset" to "top", "right", "bottom" and "left". Conversion does not work with css variables as same variable is set to all props.

Base of the new browsers list is the one React supports and is documented in HDS. Had to add some very specific rules to get inset working properly. It is supported in Safari 14.1 and Safari iOS 14.5, according to CanIUse, but those did not work.

Also had to exclude some exotic browsers  which HDS does not support.

HDS supports only versions newer than 6 months and new rule fulfills that